intermediate commit

+++ src/renderer/opengl/base/gl_framebuffer.cpp
@@ -34,7 +34,7 @@ void GLFramebuffer::upload(GraphicsContext& graphicsContext, const sp<Uploader>&
         _texture->upload(graphicsContext, nullptr);
 
     glBindFramebuffer(GL_FRAMEBUFFER, _id);
-    glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, _texture->id(), 0);
+    glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, static_cast<GLuint>(_texture->id()), 0);
     const GLenum drawBuffers[1] = {GL_COLOR_ATTACHMENT0};
     glDrawBuffers(1, drawBuffers);
     if(glCheckFramebufferStatus(GL_FRAMEBUFFER) != GL_FRAMEBUFFER_COMPLETE)
